Add gitlab cicd to gitlab/cicd
